requirements.txt
Django
psycopg2
django-model-utils==3.1.2

